requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
weasyprint>=60.0
jinja2>=3.1.2
aiosqlite>=0.19.0
httpx>=0.26.0
slowapi>=0.1.9
python-multipart>=0.0.6
stripe>=7.0.0

